+++ packages/next/server/request-meta.ts
@@ -2,8 +2,7 @@
 import type { IncomingMessage } from 'http'
 import type { ParsedUrlQuery } from 'querystring'
 import type { UrlWithParsedQuery } from 'url'
-
-import { BaseNextRequest } from './base-http'
+import type { BaseNextRequest } from './base-http'
 
 export const NEXT_REQUEST_META = Symbol('NextRequestMeta')
